grpc: Specify WithNoProxy with Unix addresses
There is a known bug [1] in grpc-go that attempts to proxy connections
destined for a unix:// prefixed address if the `http_proxy` or
`https_proxy` environment variables are set. This is undesirable
behaviour and may break all-in-one installations of Gitaly.

We already detect the connection type within client.New, so we can
explicitly provide the grpc.WithProxy dial option if we've gotten a
Unix address.

[1] https://github.com/grpc/grpc-go/issues/8207