test: Use client.New where possible
All gRPC connections outside of the tests are created via the
client.New constructor which was recently reworked in [1] to call
grpc.NewClient instead of the deprecated grpc.Dial.

Many tests however continue to call grpc.Dial, circumventing much of the
logic that's present in client.New. Update tests to use client.New
where possible, to ensure that the tests exercise the same code paths
as production.

A few notes:

- client.New performs validation on the address and expects a scheme to
  be present. Some tests are configured using Unix sockets, while others
  listen on a TCP address. These were updated to include the scheme
  prefix where necessary.
- insecure.NewCredentials() is the default when no credentials are
  specified, so that explicit dial option has been removed from the
  tests.
- Importing client.New into resolver_test.go introduces an import cycle,
  so it was not possible to update that test to use client.New. Instead,
  the test now calls the non-deprecated grpc.NewClient constructor.
- The client package exposes a dedicated DialOption for configuring a
  handshaker. This is used where required, instead of passing the
  handshaker's credentials as a grpc.DialOption.

[1] https://gitlab.com/gitlab-org/gitaly/-/merge_requests/7673