test: Generate and set TLS cert once only
After extensive debugging, it was discovered that subtle differences
between how TLS root certificates are handled on macOS and Linux meant
that it was impossible to correctly execute multiple TLS-enabled test
cases if they run in the same test process.

The auth tests configured individual TLS certificates for each test that
needed them, and changed the SSL_CERT_FILE environment variable to point
to the current certificate. On macOS [1], we manually read the
SSL_CERT_FILE and SSL_CERT_DIR environment variables and append their
contents to the system pool. On Linux [1], we delegate the functionality
to the Go standard library.

The standard library is designed to be more efficient by only reading
the environment variables once [3] per process, with an expectation that
certificates aren't going to be swapped during the lifetime of a
program. In the case of our table test, this means that any changes to
the environment variables won't get refreshed into the system pool.

Fix this by creating the certificate and setting SSL_CERT_FILE once
before the table tests are executed. Wrapping the functionality in a
once.Do() doesn't seem to work correctly, so we have to settle for this
for now. Add a second TLS-enabled test to prove that this works on macOS
and Linux.

[1] https://gitlab.com/gitlab-org/gitaly/-/blob/4d63c8f0bfb5874ba2a9a68be0b3d73d9669cd09/internal/x509/pool_darwin.go
[2] https://gitlab.com/gitlab-org/gitaly/-/blob/4d63c8f0bfb5874ba2a9a68be0b3d73d9669cd09/internal/x509/pool.go
[3] https://github.com/golang/go/blob/release-branch.go1.23/src/crypto/x509/cert_pool.go?name=release#L117-L123