diff: Refactor discarding of unused patch lines
When parsing patch output, some lines are not included with the returned
patch and discarded by calling `parser.consumeLine()`. This function has
the option to update the line stats recorded by the parser, but it is
never used.

Simplify the operation to only read the line to discard it and rename
the function to `discardLine()`.