BundleURI: Adjust configuration
This commit adjust the bundle-uri manager configuration
so that the maxBundleAge is 24 hours instead of 30 min.

This new configuration will reduce the potential
number of bundles generated for the same repo in
a given day, which will reduce load on the Gitaly
server. After some testing, it was deemed that 30
minutes was too agressive.

As a result of the test mentionned above, this commit
also adjust the interval from 30 seconds to 1 minute.

This change is needed now because this is the config
we want to go to prod with.