diff: Decouple parser state from `consumeChunkLine()`
For patch output lines that have the potential to be very long,
`consumeChunkLine()` is used to read the line in chunks. This function
depends on and modifies diff parser state which makes the parsing more
complex than it needs to be.

Refactor the function and its call sites to no longer implicitly depend
on parser state.