middleware: Extract 'peekedStream' from 'storagemgr'
The 'storagemgr' package implements a 'peekedStream' type which allows
us to peek into the first message of a stream and still allowing the
downstream users to re-read the first message. This allows the
'storagemgr' to read into important details which is only present in the
first message. In the next commit, we also need this functionality to
implement the oppurtunistic migrator for reftables. Extract this code to
the 'middleware' package so it can be re-used.