Run Praefect's subcommand tests against the actual binary
Praefect's subcommand tests are not actually running against the
real binary but rather using custom wiring to setup the tests. Running
against the real binary allows us to assert the behavior of the entire
binary which leaves little room for behavior changes to go unnoticed.

Convert the subcommand tests to run against the actual binary.