Log Praefect's error without timestamp
Praefect is currently using the 'log' package to log the error from
running the command. This includes a timestamp in the error message.
The timestamp is not particularly useful, and it's not common across
other messages logged out by Praefect either. Just log the error
out by itself as this makes it easier to assert the output of the
binary without having to adapt all of the tests.