Go-mod validator: Support replace directive
The current go-mod validator does not support
`replace` directive, only `require` ones. This
is problematic when a new dependency (A) requires
dependency (B) at version x.y.z, but we already
pinned the version of (B) to x.x.y.

The current change adds support for the `replace`
directive, which enables us to add the pin version
comment on the replace directive. This unlocks
the issue mentionned above.