service/diff: Split out blob diff logic
The DiffBlobs RPC computes and streams back diffs for each requested
blob pair. In a subsequent commit, the RPC interface is extended to use
git-diff-pairs(1) and new diff parser to compute diffs.

In preparation for this change, split out the logic only revelant to the
git-diff(1) based blob diffing to a separate function.