service/diff: Integrate git-diff-pairs(1) in DiffBlobs
The DiffBlobs RPC computes diffs for a requested set of file pairs.
Until the introduction of git-diff-pairs(1), Git did not have a way to
compute batches for file pair diffs. This required DiffBlobs to use a
separate git-diff(1) process to generate each request blob diff.

Integrate git-diff-pairs(1) into the DiffBlobs RPC and use the RawInfo
field in the request to provide input. Note that git-diff(1) based
parsing still remains implmented and used if the request uses BlobPairs
to request diffs.