log: Use WaitGroup instead of done channel
Transaction manager defers the closing of log manager before it is
initialized. This Close function waits on the pruningDone channel
which is only closed at the end of log manager init stage. Which means
if log manager failed before the pruning channel is closed, cleanup
function would go in a deadlock.
To solve this issue, we could properly refactor this to defer the
close after the init is successful. But there is a simpler solution:
If we use WaitGroup instead of a done channel, this problem would not
occur as the WaitGroup would only contain a delta when the pruning
actually started.