proto/diff: Update DiffBlobsRequest to provide raw info
The DiffBlobs RPC currently accepts a set of BlobPairs in the request to
compute diff for. These blob pairs may be an blob OID or a path-scoped
revision that resolves to a blob. To efficiently compute blob diffs,
input should instead be provided in the raw diff format. This enables
git-diff-pairs(1) for batch diff generation of file pairs.

Note that the old blob pairs field is left to remain backwards
compatible. Also, because the raw diff format requires a similarity
score for renames and copies, update the ChangedPaths message to include
this info in the DiffBlobsRequest.