raft: Rename snapshot components to reflect replica ownership
This commit renames the snapshot-related types and components to align
with our "Replica" terminology. Snapshot, SnapshotFile,
SnapshotMetadata, and RaftSnapshotter are now prefixed with "Replica" to
reflect their association with the Replica component that manages them.

The snapshotter is a critical part of our Raft implementation,
responsible for capturing and transferring state between replicas. Ther
"snapshot" term is, unfortunately, occupied by WAL. In WAL context,
"snapshot" means an isolation unit so that each request has an isolated
environment for execution. While the technique is the same, the
"snapshot" term in the Raft package has a different use. Therefore, the
term "ReplicaSnapshot" reflects its usage better.