raft: Rename Storage to ReplicaLogStore
This commit addresses terminology confusion by renaming the ambiguous
"Storage" class to "ReplicaLogStore". The term "Storage" is heavily
overloaded in our codebase - it refers to repository storage units in
Gitaly, and within the same package we already have RaftEnabledStorage
which wraps the core Storage interface.

The more specific name "ReplicaLogStore" clearly communicates this
component's responsibility: managing the persistent log state for a
specific replica in a Raft group. This distinction is important as it
establishes a clearer relationship between components - ReplicaLogStore
serves a Replica.

In addition, the same package consists of RaftEnabledStorage wrapping
around the core Storage interface. This rename restores the meaning of
that name and gets rid of a confusion source.