diff: Introduce patch parser
The current diff parser relies on git-diff(1) to be run with the `--raw`
and `--patch` options. The handling of Raw input is tightly integrated
into the parser and not required when parsing diffs via
git-diff-pairs(1) because the raw input is directly provided in the
request.

Introduce a new patch parser that accepts pre-parsed raw input to parse
git-diff-pair(1) output.