housekeeping: Simplify offloadingConfig in housekeeping config
This commit refactors `offloadingConfig` and the preparation logic for
offloading to improve clarity and correctness.

Previously, the bucket prefix and original repository path were
determined in the housekeeping manager. We now moved this logic to
the transaction manager, which is more appropriate since it’s
responsible for initiating uploads and has the right context to
resolve the original repository path.
Assigning the path UUID during upload is a more natural fit.

Additionally:
- The `Filter` field has been removed from offloadingConfig, as our
current offloading implementation only targets blobs, and this can be
hardcoded.
- `cacheRoot` is now included in offloadingConfig, and the `cachePath`
is derived from it.
- `SinkURL` has been renamed to `SinkBaseURL` to clarify that it is
not used directly to initialize the sink.