test: Add catfile test with quarantine directory
None of the catfile tests that add object readers currently configure a
quarantine repository. In practice though, quarantines are usually
applied, an example being the `ListConflictFiles` RPC. When this
happens, the Git command execution is augmented with two environment
variables:

- GIT_OBJECT_DIRECTORY which points to the new temporary quarantine
  directory.
- GIT_ALTERNATE_OBJECT_DIRECTORIES which points to the canonical
  `objects` directory.

Additionally, the cache key is also computed differently by including
the values of these variables.

Add a test that configures a quarantine repository and asserts that
the cache key is computed correctly, and that the object reader works as
expected.