command: Ignore quarantine-related cat-file errors
When a cat-file process is evicted from the catfile cache, it's possible
that it was initially spawned against a quarantine repository. When the
quarantine is eventually migrated or cleaned up [1], the underlying Git
invocation becomes invalid. Closing such a Git process causes a
"fatal: not a git repository" error due to the GIT_OBJECT_DIRECTORY not
existing anymore.

Detect and ignore these errors by introducing a new Command option that
can be configured to filter out error logs. The option sets up a
callback that's invoked before logs are produced for errored commands,
and is wired up specifically for the object reader.

[1] https://gitlab.com/gitlab-org/gitaly/blob/e360c8a785523b35ff2bced31846ac40befe5426/internal/tempdir/tempdir.go#L43