Generate help text automatically for subcommands
HideHelpCommand was set on many of the subcommands as a workaround
around a data race in the urfave/cli library. The downside of this
is that the subcommand help text is no longer automatically generated,
and hence subcommands have been documented manually in the description
field. This is error prone and bound to get out of date.

Since then, the tests have been updated to run the tests against a
real binary. This has addressed the data races in the tests as we're
no longer using the library concurrently withing a single process.
As the race has been addressed, we can now remove the recommendation
and the usage of the HideHelpCommand to again automatically generate
subcommand help text as expected.

Remove the HideHelpCommand everywhere, and the documentation advising
to use it. Adjust the description of the commands that were manually
documented the subcommands.