gitpipe: Introduce option to filter git-cat-file(1) by object type
We have recently upstreamed a new `git cat-file --filter=` option that
allows us to filter objects for the batched modes by certain criteria.
Most importantly, this new filter allows us to filter objects by their
respective type.

When used like this, git-cat-file(1) knows to consult packfile bitmaps
to locate such objects. This fixes a significant performance issue that
we have been observing: in order to yield all objects of a specific
type, we always had to read through _all_ objects only to find those
that we are interested in. But packfile bitmaps already encode required
information of where objects of a specific type are located. As such,
iterating through objects with bitmaps doesn't scale linearly with the
total number of objects anymore, but instead with the number of objects
of a specific type.

This can make a significant difference depending on the object type that
one wants to read. The following benchmark for example yields commits in
the Chromium repository with and without bitmaps:

    Benchmark 1: cat-file with filter=object:type=commit (revision = HEAD~)
      Time (mean ± σ):     86.444 s ±  4.081 s    [User: 36.830 s, System: 11.312 s]
      Range (min … max):   80.305 s … 93.104 s    10 runs

    Benchmark 2: cat-file with filter=object:type=commit (revision = HEAD)
      Time (mean ± σ):      2.089 s ±  0.015 s    [User: 1.872 s, System: 0.207 s]
      Range (min … max):    2.073 s …  2.119 s    10 runs

    Summary
      cat-file with filter=object:type=commit (revision = HEAD) ran
       41.38 ± 1.98 times faster than cat-file with filter=object:type=commit (revision = HEAD~)

Introduce a new option that allows callers to specify such object
filters. This option is only to be used with `CatfileInfoAllObjects()`,
as it does not make much sense for any of the user uses. Furthermore, it
can only be used with Git v2.49.0.gl2 or later, which contains backports
of the respective upstream commits.

This option will be used in subsequent commits to speed up a couple of
RPCs that use the `--batch-all-objects` mode.