raft: Add membership management operations to RaftReplica
This commit adds methods to manage Raft cluster membership, additionally
we removed the test TestReplica_ProcessConfChange as we will no longer
submit the configuration change directly. In this iteration we didn't
add the method to promote a learner node, promotion requires verifying
that the learner is fully in sync with the leader. We'll introduce this
method in a future iteration once transport is enabled.