Consume stdin from mock reference transaction hook
TestPostReceivePack_notAllowed is flaky as the mock hook manager
used in the test does not consume the reference transaction hook
input before returning successfully. This can lead to gitaly-hooks
failing as the RPC stream finished with a successful response before
the whole stdin has been written to the stream. This leads to
subsequent writes on the stream failing.

Fix the flake by consuming the stdin stream from the mock hook as
the real one would.