Skip the quarantined test directly
SkipQuarantinedTest is taking in the quarantined test names as
parameters and skipping the test if the name matches. This has led
to issues where we have quarantine instructions against non-existing
tests. It's better to invoke the quarantine directly on the running
test. This way the quarantining doesn't depend on the test name which
may be changed.

The context value is moved below the skip in TestRemoveRepository as
the linter now notices it is unused and will complain about it.