log: Remove log.AddFields
Gitaly runs with a single instance of a log.Logger, which is typically
either injected into context or passed as function arguments. The
logger's fields are part of this singleton, and get manipulated in
several places:

- log.FromContext which extracts a logger from context, performing a
  read on the fields map.
- log.AddFields which adds fields to the logger inside the context,
  performing a write on the fields map.
- LogrusLogger.log which is similar to log.FromContext, performing a
  read on the fields map.

It's possible for concurrent goroutines to read and write to the fields
map through these functions, causing a
"concurrent map iteration and map write" error. While logrus uses a
mutex internally to synchronise writing logs, it doesn't for the fields
map.

AddFields can be replaced with calls to CustomFieldsFromContext and
RecordMetadata, which already have a mutex to synchronise access to the
fields. This also helps in our effort [1] to remove logrus and
ctxlogrus.

[1] https://gitlab.com/gitlab-org/gitaly/-/work_items/5703