Synchronize negotiation timeout test properly
TestUploadPackWithSidechannel_timeout attempst to test the negotiation
timeout on upload pack. The test case is not properly synchronized which
leads to flakiness.

Once the client receives the sidechannel connection, it waits reads a
channel to see whether the server has created negotiation timeout, and
proceeds to close the connection. It's not guaranteed the server has
actually observed the timeout yet. As the client closes the connection,
the server sees the connection being abruptly broken and bubbles up
the error. If we want to simulate a negotiation timeout, the client
shouldn't hang up on the server but should instead block until the
server hangs up due to the timeout.

Fix the flake by blocking in the client until the server hangs up, and
remove the broken synchronization logic.