Bundle-URI custom gitlab-runner
Bundle-URI is a feature not yet available in production
because it requires a specific version of gitlab-runner
which is not tagged yet, and Git version 2.49, which is
not available in the current `helper-image`.

This commit introduces a new script to create a custom
gitlab-runner fleet that will use a built-from-source
version of gitlab-runner using a specific revision that
contains the changes needed. It also configures the
runner to use a custom-built `image-helper` that runs
Git version 2.49.