Fix subprocesses leaking when failing to configure its cgroup
Gitaly is leaking child processes from command.New when they fail
to be added to a cgroup after they've been started. If AddCommand
fails, we'll return a nil command even though the command was actually
started. Fix this issue by tearing down the command if an error
occurs after the process is started.

We have to adjust also the subprocess logger's clean up routine around
this change. We have to close the pipes on both failures before the
process is started, and the write end expicitly if the process fails
to be started after the log processing goroutine has been started.