Bundle-URI: Use custom-gitlab-runner in cicd build
Bundle-URI is now available in production but is still
fairly new and has not been tested with real live traffic.
To avoid enabling it widely and suffer unexpected errors, we
enabled it only for the Gitaly project; that way we can
test our own feature with our own project.

This commit adds the necessary steps in the .gitlab-ci.yaml
file to use a custom-built gitlab-runner with bundle-uri
enabled.