reftable: Retry the RPC multiple times in tests
In the TestUnaryInterceptor and TestStreamInterceptor tests, we initiate
a migration indirectly by calling an RPC which would trigger the
`RegisterMigration` function.

This `RegisterMigration` function is a non-blocking function which
attempts to register a migration. Due to its nature, it is completely
possible that it doesn't register a migration. So let's attempt to do
this 3 times to increase the chances of this succeeding.