test: Refactor command tests to use logrus.Entry
The TestCommand_stderr tests currently use a combination of log level,
log output, and stderr length to perform assertions on the produced
logs. The length is primarily used in test cases that ensure the various
stderr length limits are correctly enforced.

We have different conditions that guard what gets asserted based on the
fields provided in the test expectations. This can get a bit confusing
to read, and a subsequent change will require us to add yet another
field to assert that no logs get produced.

Refactor the tests to bundle the expected values within *logrus.Entry.
Instead of asserting stderr length, we can just compute the content and
assert the output directly.