storagemgr: Remove partitions from closing list
Once a partition is evicted from inactive partitions list or it has
returned from it's Run function due to an error, the partition still
needs to wait for the in-flight transactions to finish before it can
completely shutdown. It is moved into the closingPartitions map to
keep a handle on the partition so it can be waited on by the
StorageManager to properly clean up when closing the storage.

There is a problem with this logic as we are never dropping the
partitions from the closingPartitions map. This leads to a memory
leak as the reference to the partition prevents it from being garbage
collected. There was also a bug where the partition was being added
to the closingPartitions list twice redundantly. With this change we
are removing the second case where it is added to the closingPartitions
list and instead we remove it from the list.