housekeeping: Record multi-pack-index update in transaction
After a repacking operation we save record the file deletion and
creation operations in WAL. This is done simply by storing the list
of packfiles before and after repacking and comparing them by name.
But this comparison doesn't work for multi-pack-index file as it's
name does not change. Therefore we need to record the operation for
midx file explicitly by comparing it's modification time.