housekeeping: Prune empty directories
Housekeeping with transactions removes loose reference directories
as part of repacking. However, it only removes the directories that
had loose references in them during the reference packing run.
Empty directories that didn't contain any loose references to pack
are not removed. This leads to empty directories never being pruned.

With this change we will also add empty directories to the list of
paths to be pruned, so that they will get deleted in verifyPackRefsFiles