housekeeping: Add scheduling middleware
We are not currently performing repository housekeeping often enough,
leading to repositories accumulating hundreds of packfiles before they
get repacked. This leads to performance degradation, especially when
transactions and the WAL are enabled.

Housekeeping was historically scheduled in two places:
- By Rails, after a specific (200) number of pushes was performed
  against a repository. Once this threshold was reached, Rails would
  invoke the OptimizeRepository RPC.
- In Gitaly's DailyOptimizationWorker, which randomly walks storages and
  executes the OptimizeRepository function on it. This does not ensure
  full coverage as the worker only executes for a certain duration.

Adding scheduling middleware is the first step to re-integrating
housekeeping tasks into Gitaly, removing reliance on Rails. The
middleware is transaction-independent and intercepts all gRPC requests.
It ensures that at most one invocation of OptimizeRepository runs for
each repository at a given time, and this applies to housekeeping tasks
it starts itself as well as those requested via RPC.