Log the commit LSN of a write transaction
We currently do not log the commit LSN of a transaction. This makes
it difficult to correlate transactions that conflicted with each
other. On conflicts, we log the items that conflicted along with their
LSN. Logging the commit LSN of writes will make it easier to identify
which exact write produced the concurrent change that conflicted.

The first solution that comes to mind is to simply include the commit
LSN to the fields in the context and log it as a part of the RPC status
log line. This doesn't work well as we have RPCs that run multiple
transactions, and we run transactions also outside of the context of
RPCs for example in the daily maintenance jobs.

This commit returns the commit LSN of a transaction from the Commit()
call. This way the caller can decide what to do with it. In this case,
we log the transaction commits where they happen along with the commit
LSN of the transaction.