recovery: Add new subcmd to recover partitions
Previouly we have developed BackupPartition RPC which archives
the fs state of a partition to configured backup sink. Now we
are adding a command to restore a partition from the archive
which will be used for restoring partitions to a fresh gitaly
instance.
The command also handles restoring the KV state from the same
tar archive.