protobuf: Add message_regex field to FindCommitsRequest
Add a new field to the FindCommitsRequest proto message to allow filtering
commits by their message content using a regular expression pattern.

The field is being added to provide a more user-friendly search
experience, and is part of an effort [1] to improve the Commit List page
with additional abilities.

1. https://gitlab.com/gitlab-org/gitlab/-/issues/542319

Changelog: added