ci: Set OVERRIDE_GIT_VERSION for nightly tests
The OVERRIDE_GIT_VERSION variable was introduced [1] to allow an
arbitrary version string to be specified. This would become the value
reported by `git --version` and subsequently parsed by Gitaly's version
checking logic. We sometimes need to check the Git version in order to
determine if a particular feature is avaialble, or change how test
assertions are made.

There was previous discussion in [1] on whether OVERRIDE_GITALY_VERSION
should be used to enable upstream feature branches to be tested. It was
decided that introducing a `.gl*` suffix to the version string would be
a better idea. We've done this in practice for the cat-file --filter
option [2] for example.

For the nightly tests which execute on the `master` and `next` branches
however, this wouldn't work. Those branches would contain the latest
patches, but the version reported by `git --version` wouldn't contain
the GitLab suffix. We can use OVERRIDE_GIT_VERSION in these cases to set
an arbitrarily high value so checks such as [2] continue to function as
intended.

[1] https://gitlab.com/gitlab-org/gitaly/-/merge_requests/6986
[2] https://gitlab.com/gitlab-org/gitaly/blob/7e449770eb53b473a336d160df516aad427244ca/internal/git/version.go#L92-L94