Lock the latest reftable in write transactions
TransactionManager currently handles logging reftables by replaying
the captured logical reference changes in the critical section and
recording the file modifications. This is inefficient, and we want
to move towards capturing the file modifications as performed during
the transaction in the snapshot. We'll take the table files, manually
resequence their update index, and merge them with the latest tables.list
file in the repository. This way we can get rid of the update-ref command
in the critical section, and the associated repository snapshot.

The problem with doing this is that currently auto-compaction may
merge the new table files produced by the transaction into the tables
that already existed in the repository before the transaction started.
If multiple transactions do this, they'd concurrently modify the existing
table files and lead produce conflicts that lead to transaction aborts.

To avoid this, we want to prevent auto-compaction from merging the new
writes into the existing tables in the repository. This way we can capture
the new table files without failing due to conflicts on concurrent
modifications to the existing files.

This commit writes a .lock file for the latest table file in the repository
when beginning a write transaction. The .lock file prevents Git from auto-
compacting the locked table file and any earlier tables. We unfortunately
need to write a file to the disk explicitly for now as Git does not provide
enough control otherwise to disable auto-compaction of tables below a given
index.

Housekeeping transactions however are expected to compact all tables. To
allow for this, the housekeeping logic has been updated to drop the lock
file before running compaction.

As we're parsing the reftable file here in a stricter manner, we also
notice that our housekeeping logic for reftables isn't writing a newline
at the end of the tables.list file as Git does. Fix this bug while at it
as there's no reason to allow varying structures to the file, and handling
the missing new line would add unnecessary complexity.

Externally visible behavior remains unchanged but transactions no longer
merge new writes into the existing tables while running.