Discern different housekeeping transactions in logs
We run multiple transactions as part of OptimizeRepository.

The top most transaction is a read-only transaction that decides
based on the heuristics whether to run housekeeping. If we should
run compaction, we do so in two different write transactions that
respectively pack references and objects. Running the operations in
these two transactions allows the faster operation to commit faster,
and reduces the chances of conflicts aborting both operations.

The downside with this is that we don't really have good knowledge from
just the status log of the OptimizeRepository call on when these
transactions commit. This makes it difficult to correlate the changes
the housekeeping transaction did back to them from the logs of other
transactions that aborted on them.

Pass a description into the log statement with each of the transaction
that allows us to tell them apart. This way we'll have easier time
attributing conflicts to a certain housekeeping operation.