Pass log fields through HookPayload to hooks
We're currently not passing the log fields through to the hooks.
This means any logs written in the hooks won't carry correlation id
or any other useful information. This commit passes the log fields
through the HookPayloads to the hook service calls so they can
include the expected context in their log statements.

We'll soon use this to log transaction commits in post-receive hook
with full context.