housekeeping/manager: Rename TestPackRefsIfNeeded to TestPackRefsInhibitors
The test 'TestPackRefsIfNeeded' is used to verify the inhibitor logic,
which blocks multiple 'git pack-refs' from running at the same time. As
such its naming is incorrect, so let's rename it to
'TestPackRefsInhibitors'.

The test also checks specifically for the 'packed-refs' file, which is:
1. Not the purpose of the test. It shouldn't be checking if
'git-pack-refs(1)' really did its job.
2. Is specific to the files backend, making this test only work with the
files backend.

So let's clean that up. Also while we're here, let's modify the loops to
use the shiny new syntax which the LSP is suggesting.