Assert the produced output in TestListStoragesSubcommand
TestListStoragesSubcommand is producing the expected output by
duplicating the table writing code from the actual subcommand. This
is not ideal as first of all it wouldn't notice any issues in the
output since the output is produced by the same code as in the
actual subcommand. Secondly, it relies on the tablewriter package's
API which is about to be changed significantly in an update, and this
would break our tests. Adapting the tests with the change makes it
difficult to assert the output remains the same.

Instead of generating the expected output using tablewriter, simply
hard code the expected output so we'll know if it changes.