housekeeping: Add repacking heuristics for reftables
The 'HeuristicalOptimizationStrategy' uses heuristics to decide if
repacking should take place or not. This is done via the
'ShouldRepackReferences' function.

For the files backend, the function uses a logarithmic scale to allow a
certain number of loose refs to packed-refs. We cannot use the same
strategy in reftables. In reftables, we mimic the strategy used in the
auto-compaction code of git. Which is, to only repack the tables if they
do not follow a geometric progression.